requirements.txt
lxml
Pillow
numpy
wheel
requests
